+++ src/utils/helpers/index.ts
@@ -1,3 +1,2 @@
 export * from './auth';
-export * from './axiosHandler';
 export * from './error';
